split programming exercise service into multiple classes
add test for programming exercise export
remove old code (create programming exercise / edit link) is not necessary any more
additional logic refactorings
