+++ src/main/java/de/tum/in/www1/artemis/service/ProgrammingExerciseService.java
@@ -10,59 +10,32 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.ZonedDateTime;
 import java.util.*;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
 
-import javax.validation.constraints.NotNull;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
-import javax.xml.xpath.XPath;
-import javax.xml.xpath.XPathConstants;
-import javax.xml.xpath.XPathException;
-import javax.xml.xpath.XPathFactory;
+import javax.annotation.Nullable;
 
 import org.apache.commons.lang3.ArrayUtils;
-import org.apache.http.HttpException;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.core.io.support.ResourcePatternUtils;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.util.Pair;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.w3c.dom.Document;
-import org.w3c.dom.Node;
-import org.xml.sax.InputSource;
-import org.xml.sax.SAXException;
 
 import de.tum.in.www1.artemis.domain.*;
 import de.tum.in.www1.artemis.domain.enumeration.*;
 import de.tum.in.www1.artemis.domain.participation.*;
-import de.tum.in.www1.artemis.exception.GitException;
 import de.tum.in.www1.artemis.repository.*;
 import de.tum.in.www1.artemis.service.connectors.CIPermission;
 import de.tum.in.www1.artemis.service.connectors.ContinuousIntegrationService;
 import de.tum.in.www1.artemis.service.connectors.GitService;
 import de.tum.in.www1.artemis.service.connectors.VersionControlService;
+import de.tum.in.www1.artemis.service.scheduled.ProgrammingExerciseScheduleService;
 import de.tum.in.www1.artemis.service.util.structureoraclegenerator.OracleGenerator;
 import de.tum.in.www1.artemis.web.rest.dto.PageableSearchDTO;
-import de.tum.in.www1.artemis.web.rest.dto.RepositoryExportOptionsDTO;
 import de.tum.in.www1.artemis.web.rest.dto.SearchResultPageDTO;
 import de.tum.in.www1.artemis.web.rest.errors.EntityNotFoundException;
 
@@ -77,13 +50,9 @@ public class ProgrammingExerciseService {
 
     private final GitService gitService;
 
-    private final ExerciseHintService exerciseHintService;
-
     private final Optional<VersionControlService> versionControlService;
 
     private final Optional<ContinuousIntegrationService> continuousIntegrationService;
-
-    private final ProgrammingExerciseParticipationService programmingExerciseParticipationService;
 
     private final TemplateProgrammingExerciseParticipationRepository templateProgrammingExerciseParticipationRepository;
 
@@ -97,26 +66,23 @@ public class ProgrammingExerciseService {
 
     private final AuthorizationCheckService authCheckService;
 
+    private final ProgrammingExerciseScheduleService programmingExerciseScheduleService;
+
+    private final GroupNotificationService groupNotificationService;
+
     private final ResourceLoader resourceLoader;
 
-    private final ProgrammingExerciseTestCaseRepository programmingExerciseTestCaseRepository;
-
-    @Value("${artemis.repo-download-clone-path}")
-    private String REPO_DOWNLOAD_CLONE_PATH;
-
     public ProgrammingExerciseService(ProgrammingExerciseRepository programmingExerciseRepository, FileService fileService, GitService gitService,
-            ExerciseHintService exerciseHintService, Optional<VersionControlService> versionControlService, Optional<ContinuousIntegrationService> continuousIntegrationService,
-            ProgrammingExerciseParticipationService programmingExerciseParticipationService, ProgrammingExerciseTestCaseRepository programmingExerciseTestCaseRepository,
+            Optional<VersionControlService> versionControlService, Optional<ContinuousIntegrationService> continuousIntegrationService,
             TemplateProgrammingExerciseParticipationRepository templateProgrammingExerciseParticipationRepository,
             SolutionProgrammingExerciseParticipationRepository solutionProgrammingExerciseParticipationRepository, ParticipationService participationService,
-            ResultRepository resultRepository, UserService userService, AuthorizationCheckService authCheckService, ResourceLoader resourceLoader) {
+            ResultRepository resultRepository, UserService userService, AuthorizationCheckService authCheckService, ResourceLoader resourceLoader,
+            ProgrammingExerciseScheduleService programmingExerciseScheduleService, GroupNotificationService groupNotificationService) {
         this.programmingExerciseRepository = programmingExerciseRepository;
         this.fileService = fileService;
         this.gitService = gitService;
-        this.exerciseHintService = exerciseHintService;
         this.versionControlService = versionControlService;
         this.continuousIntegrationService = continuousIntegrationService;
-        this.programmingExerciseParticipationService = programmingExerciseParticipationService;
         this.templateProgrammingExerciseParticipationRepository = templateProgrammingExerciseParticipationRepository;
         this.solutionProgrammingExerciseParticipationRepository = solutionProgrammingExerciseParticipationRepository;
         this.participationService = participationService;
@@ -124,122 +90,8 @@ public class ProgrammingExerciseService {
         this.userService = userService;
         this.authCheckService = authCheckService;
         this.resourceLoader = resourceLoader;
-        this.programmingExerciseTestCaseRepository = programmingExerciseTestCaseRepository;
-    }
-
-    /**
-     * Adds the student id of the given student participation to the project name in all .project (Eclipse)
-     * and pom.xml (Maven) files found in the given repository.
-     *
-     * @param repo The repository for which the student id should get added
-     * @param programmingExercise The checked out exercise in the repository
-     * @param participation The student participation for the student id, which should be added.
-     */
-    public void addStudentIdToProjectName(Repository repo, ProgrammingExercise programmingExercise, StudentParticipation participation) {
-        String studentId = participation.getStudent().getLogin();
-
-        // Get all files in repository expect .git files
-        List<String> allRepoFiles = listAllFilesInPath(repo.getLocalPath());
-
-        // is Java programming language
-        if (programmingExercise.getProgrammingLanguage() == ProgrammingLanguage.JAVA) {
-            // Filter all Eclipse .project files
-            List<String> eclipseProjectFiles = allRepoFiles.stream().filter(file -> file.endsWith(".project")).collect(Collectors.toList());
-
-            for (String eclipseProjectFilePath : eclipseProjectFiles) {
-                File eclipseProjectFile = new File(eclipseProjectFilePath);
-                // Check if file exists and full file name is .project and not just the file ending.
-                if (!eclipseProjectFile.exists() || !eclipseProjectFile.getName().equals(".project")) {
-                    continue;
-                }
-
-                try {
-                    // 1- Build the doc from the XML file
-                    Document doc = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(new InputSource(eclipseProjectFile.getPath()));
-                    doc.setXmlStandalone(true);
-
-                    // 2- Find the node with xpath
-                    XPath xPath = XPathFactory.newInstance().newXPath();
-                    Node nameNode = (Node) xPath.compile("/projectDescription/name").evaluate(doc, XPathConstants.NODE);
-
-                    // 3- Append Student Id to Project Name
-                    if (nameNode != null) {
-                        nameNode.setTextContent(nameNode.getTextContent() + " " + studentId);
-                    }
-
-                    // 4- Save the result to a new XML doc
-                    Transformer xformer = TransformerFactory.newInstance().newTransformer();
-                    xformer.transform(new DOMSource(doc), new StreamResult(new File(eclipseProjectFile.getPath())));
-
-                }
-                catch (SAXException | IOException | ParserConfigurationException | TransformerException | XPathException ex) {
-                    log.error("Cannot rename .project file in " + repo.getLocalPath() + " due to the following exception: " + ex);
-                }
-            }
-
-            // Filter all pom.xml files
-            List<String> pomFiles = allRepoFiles.stream().filter(file -> file.endsWith("pom.xml")).collect(Collectors.toList());
-            for (String pomFilePath : pomFiles) {
-                File pomFile = new File(pomFilePath);
-                // check if file exists and full file name is pom.xml and not just the file ending.
-                if (!pomFile.exists() || !pomFile.getName().equals("pom.xml")) {
-                    continue;
-                }
-
-                try {
-                    // 1- Build the doc from the XML file
-                    Document doc = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(new InputSource(pomFile.getPath()));
-                    doc.setXmlStandalone(true);
-
-                    // 2- Find the relevant nodes with xpath
-                    XPath xPath = XPathFactory.newInstance().newXPath();
-                    Node nameNode = (Node) xPath.compile("/project/name").evaluate(doc, XPathConstants.NODE);
-                    Node artifactIdNode = (Node) xPath.compile("/project/artifactId").evaluate(doc, XPathConstants.NODE);
-
-                    // 3- Append Student Id to Project Names
-                    if (nameNode != null) {
-                        nameNode.setTextContent(nameNode.getTextContent() + " " + studentId);
-                    }
-                    if (artifactIdNode != null) {
-                        String artifactId = (artifactIdNode.getTextContent() + "-" + studentId).replaceAll(" ", "-").toLowerCase();
-                        artifactIdNode.setTextContent(artifactId);
-                    }
-
-                    // 4- Save the result to a new XML doc
-                    Transformer xformer = TransformerFactory.newInstance().newTransformer();
-                    xformer.transform(new DOMSource(doc), new StreamResult(new File(pomFile.getPath())));
-
-                }
-                catch (SAXException | IOException | ParserConfigurationException | TransformerException | XPathException ex) {
-                    log.error("Cannot rename pom.xml file in " + repo.getLocalPath() + " due to the following exception: " + ex);
-                }
-            }
-        }
-
-        try {
-            gitService.stageAllChanges(repo);
-            gitService.commit(repo, "Add Student Id to Project Name");
-        }
-        catch (GitAPIException ex) {
-            log.error("Cannot stage or commit to the repo " + repo.getLocalPath() + " due to the following exception: " + ex);
-        }
-    }
-
-    /**
-     * Get all files in path except .git files
-     *
-     * @param path The path for which all file names should be listed
-     * @return A list of all file names under the given path
-     */
-    private List<String> listAllFilesInPath(Path path) {
-        List<String> allRepoFiles = null;
-        try (Stream<Path> walk = Files.walk(path)) {
-            allRepoFiles = walk.filter(Files::isRegularFile).map(Path::toString).filter(s -> !s.contains(".git")).collect(Collectors.toList());
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-        return allRepoFiles;
+        this.programmingExerciseScheduleService = programmingExerciseScheduleService;
+        this.groupNotificationService = groupNotificationService;
     }
 
     // TODO We too many many generic throws Exception declarations.
@@ -296,8 +148,8 @@ public class ProgrammingExerciseService {
         // Save participations to get the ids required for the webhooks
         templateParticipation.setProgrammingExercise(programmingExercise);
         solutionParticipation.setProgrammingExercise(programmingExercise);
-        templateParticipation = templateProgrammingExerciseParticipationRepository.save(templateParticipation);
-        solutionParticipation = solutionProgrammingExerciseParticipationRepository.save(solutionParticipation);
+        templateProgrammingExerciseParticipationRepository.save(templateParticipation);
+        solutionProgrammingExerciseParticipationRepository.save(solutionParticipation);
 
         URL exerciseRepoUrl = versionControlService.get().getCloneRepositoryUrl(projectKey, exerciseRepoName).getURL();
         URL testsRepoUrl = versionControlService.get().getCloneRepositoryUrl(projectKey, testRepoName).getURL();
@@ -354,7 +206,28 @@ public class ProgrammingExerciseService {
         // not yet saved in the database, so we cannot save the submission accordingly (see ProgrammingSubmissionService.notifyPush)
         versionControlService.get().addWebHooksForExercise(programmingExercise);
 
+        programmingExerciseScheduleService.scheduleExerciseIfRequired(programmingExercise);
+        groupNotificationService.notifyTutorGroupAboutExerciseCreated(programmingExercise);
+
         return programmingExercise;
+    }
+
+    /**
+     *
+     * @param programmingExercise the changed programming exercise with its new values
+     * @param notificationText optional text about the changes for a notification
+     * @return the updates programming exercise from the database
+     */
+    public ProgrammingExercise updateProgrammingExercise(ProgrammingExercise programmingExercise, @Nullable String notificationText) {
+        ProgrammingExercise savedProgrammingExercise = programmingExerciseRepository.save(programmingExercise);
+
+        // TODO: should the call `scheduleExerciseIfRequired` not be moved into the service?
+        programmingExerciseScheduleService.scheduleExerciseIfRequired(savedProgrammingExercise);
+        if (notificationText != null) {
+            groupNotificationService.notifyStudentGroupAboutExerciseUpdate(savedProgrammingExercise, notificationText);
+        }
+
+        return savedProgrammingExercise;
     }
 
     /**
@@ -531,23 +404,13 @@ public class ProgrammingExerciseService {
     }
 
     /**
-     * Find the ProgrammingExercise where the given Participation is the solution or template Participation
-     *
-     * @param participation The solution or template participation
-     * @return The ProgrammingExercise where the given Participation is the solution or template Participation
-     */
-    public Optional<ProgrammingExercise> getExerciseForSolutionOrTemplateParticipation(Participation participation) {
-        return programmingExerciseRepository.findOneByTemplateParticipationIdOrSolutionParticipationId(participation.getId());
-    }
-
-    /**
      * Find a programming exercise by its id.
      * 
      * @param programmingExerciseId of the programming exercise.
      * @return The programming exercise related to the given id
      * @throws EntityNotFoundException the programming exercise could not be found.
      */
-    public ProgrammingExercise findById(Long programmingExerciseId) throws EntityNotFoundException {
+    public ProgrammingExercise findWithTemplateParticipationAndSolutionParticipationById(Long programmingExerciseId) throws EntityNotFoundException {
         Optional<ProgrammingExercise> programmingExercise = programmingExerciseRepository.findWithTemplateParticipationAndSolutionParticipationById(programmingExerciseId);
         if (programmingExercise.isPresent()) {
             return programmingExercise.get();
@@ -644,11 +507,13 @@ public class ProgrammingExerciseService {
      *
      * @param programmingExerciseId ProgrammingExercise Id.
      * @param problemStatement markdown of the problem statement.
+     * @param notificationText optional text for a notification to all students about the update
      * @return the updated ProgrammingExercise object.
      * @throws EntityNotFoundException if there is no ProgrammingExercise for the given id.
      * @throws IllegalAccessException if the user does not have permissions to access the ProgrammingExercise.
      */
-    public ProgrammingExercise updateProblemStatement(Long programmingExerciseId, String problemStatement) throws EntityNotFoundException, IllegalAccessException {
+    public ProgrammingExercise updateProblemStatement(Long programmingExerciseId, String problemStatement, @Nullable String notificationText)
+            throws EntityNotFoundException, IllegalAccessException {
         Optional<ProgrammingExercise> programmingExerciseOpt = programmingExerciseRepository.findWithTemplateParticipationAndSolutionParticipationById(programmingExerciseId);
         if (programmingExerciseOpt.isEmpty()) {
             throw new EntityNotFoundException("Programming exercise not found with id: " + programmingExerciseId);
@@ -659,7 +524,11 @@ public class ProgrammingExerciseService {
             throw new IllegalAccessException("User with login " + user.getLogin() + " is not authorized to access programming exercise with id: " + programmingExerciseId);
         }
         programmingExercise.setProblemStatement(problemStatement);
-        return programmingExerciseRepository.save(programmingExercise);
+        ProgrammingExercise updatedProgrammingExercise = programmingExerciseRepository.save(programmingExercise);
+        if (notificationText != null) {
+            groupNotificationService.notifyStudentGroupAboutExerciseUpdate(updatedProgrammingExercise, notificationText);
+        }
+        return updatedProgrammingExercise;
     }
 
     /**
@@ -799,7 +668,7 @@ public class ProgrammingExerciseService {
 
     public boolean hasAtLeastOneStudentResult(ProgrammingExercise programmingExercise) {
         // Is true if the exercise is released and has at least one result.
-        // TODO: We can't use the resultService here due to a circular dependency issue.
+        // We can't use the resultService here due to a circular dependency issue.
         return resultRepository.existsByParticipation_ExerciseId(programmingExercise.getId());
     }
 
@@ -830,118 +699,7 @@ public class ProgrammingExerciseService {
         return new SearchResultPageDTO<>(exercisePage.getContent(), exercisePage.getTotalPages());
     }
 
-    /**
-     * Imports a programming exercise creating a new entity, copying all basic values and saving it in the database.
-     * All basic include everything except for repositories, or build plans on a remote version control server, or
-     * continuous integration server. <br>
-     * There are however, a couple of things that will never get copied:
-     * <ul>
-     *     <li>The ID</li>
-     *     <li>The template and solution participation</li>
-     *     <li>The number of complaints, assessments and more feedback requests</li>
-     *     <li>The tutor/student participations</li>
-     *     <li>The questions asked by students</li>
-     *     <li>The example submissions</li>
-     * </ul>
-     *
-     * @param templateExercise The template exercise which should get imported
-     * @param newExercise The new exercise already containing values which should not get copied, i.e. overwritten
-     * @return The newly created exercise
-     */
-    @Transactional
-    public ProgrammingExercise importProgrammingExerciseBasis(final ProgrammingExercise templateExercise, final ProgrammingExercise newExercise) {
-        // Set values we don't want to copy to null
-        setupExerciseForImport(newExercise);
-        newExercise.generateAndSetProjectKey();
-        final var projectKey = newExercise.getProjectKey();
-        final var templatePlanName = BuildPlanType.TEMPLATE.getName();
-        final var solutionPlanName = BuildPlanType.SOLUTION.getName();
-
-        programmingExerciseParticipationService.setupInitialSolutionParticipation(newExercise, projectKey, solutionPlanName);
-        programmingExerciseParticipationService.setupInitalTemplateParticipation(newExercise, projectKey, templatePlanName);
-        setupTestRepository(newExercise, projectKey);
-        initParticipations(newExercise);
-
-        // Hints and test cases
-        exerciseHintService.copyExerciseHints(templateExercise, newExercise);
-        programmingExerciseRepository.save(newExercise);
-        importTestCases(templateExercise, newExercise);
-
-        return newExercise;
-    }
-
-    /**
-     * Import all base repositories from one exercise. These include the template, the solution and the test
-     * repository. Participation repositories from students or tutors will not get copied!
-     *
-     * @param templateExercise The template exercise having a reference to all base repositories
-     * @param newExercise The new exercise without any repositories
-     */
-    public void importRepositories(final ProgrammingExercise templateExercise, final ProgrammingExercise newExercise) {
-        final var targetProjectKey = newExercise.getProjectKey();
-        final var sourceProjectKey = templateExercise.getProjectKey();
-        final var templateParticipation = newExercise.getTemplateParticipation();
-        final var solutionParticipation = newExercise.getSolutionParticipation();
-
-        // First, create a new project for our imported exercise
-        versionControlService.get().createProjectForExercise(newExercise);
-        // Copy all repositories
-        final var reposToCopy = List.of(Pair.of(RepositoryType.TEMPLATE, templateExercise.getTemplateRepositoryName()),
-                Pair.of(RepositoryType.SOLUTION, templateExercise.getSolutionRepositoryName()), Pair.of(RepositoryType.TESTS, templateExercise.getTestRepositoryName()));
-        reposToCopy.forEach(repo -> versionControlService.get().copyRepository(sourceProjectKey, repo.getSecond(), targetProjectKey, repo.getFirst().getName()));
-        // Add the necessary hooks notifying Artemis about changes after commits have been pushed
-        versionControlService.get().addWebHooksForExercise(newExercise);
-    }
-
-    /**
-     * Imports all base build plans for an exercise. These include the template and the solution build plan, <b>not</b>
-     * any participation plans!
-     *
-     * @param templateExercise The template exercise which plans should get copied
-     * @param newExercise The new exercise to which all plans should get copied
-     * @throws HttpException If the copied build plans could not get triggered
-     */
-    public void importBuildPlans(final ProgrammingExercise templateExercise, final ProgrammingExercise newExercise) throws HttpException {
-        final var templateParticipation = newExercise.getTemplateParticipation();
-        final var solutionParticipation = newExercise.getSolutionParticipation();
-        final var templatePlanName = BuildPlanType.TEMPLATE.getName();
-        final var solutionPlanName = BuildPlanType.SOLUTION.getName();
-        final var templateKey = templateExercise.getProjectKey();
-        final var targetKey = newExercise.getProjectKey();
-        final var targetName = newExercise.getCourse().getShortName().toUpperCase() + " " + newExercise.getTitle();
-        final var targetExerciseProjectKey = newExercise.getProjectKey();
-
-        // Clone all build plans, enable them and setup the initial participations, i.e. setting the correct repo URLs and
-        // running the plan for the first time
-        continuousIntegrationService.get().copyBuildPlan(templateKey, templatePlanName, targetKey, targetName, templatePlanName);
-        continuousIntegrationService.get().copyBuildPlan(templateKey, solutionPlanName, targetKey, targetName, solutionPlanName);
-        giveCIProjectPermissions(newExercise);
-        continuousIntegrationService.get().enablePlan(targetExerciseProjectKey, templateParticipation.getBuildPlanId());
-        continuousIntegrationService.get().enablePlan(targetExerciseProjectKey, solutionParticipation.getBuildPlanId());
-
-        // update 2 repositories for the template (BASE) build plan
-        continuousIntegrationService.get().updatePlanRepository(targetExerciseProjectKey, templateParticipation.getBuildPlanId(), ASSIGNMENT_REPO_NAME, targetExerciseProjectKey,
-                newExercise.getTemplateRepositoryUrl(), Optional.of(List.of(ASSIGNMENT_REPO_NAME)));
-        continuousIntegrationService.get().updatePlanRepository(targetExerciseProjectKey, templateParticipation.getBuildPlanId(), TEST_REPO_NAME, targetExerciseProjectKey,
-                newExercise.getTestRepositoryUrl(), Optional.empty());
-
-        // update 2 repositories for the solution (SOLUTION) build plan
-        continuousIntegrationService.get().updatePlanRepository(targetExerciseProjectKey, solutionParticipation.getBuildPlanId(), ASSIGNMENT_REPO_NAME, targetExerciseProjectKey,
-                newExercise.getSolutionRepositoryUrl(), Optional.empty());
-        continuousIntegrationService.get().updatePlanRepository(targetExerciseProjectKey, solutionParticipation.getBuildPlanId(), TEST_REPO_NAME, targetExerciseProjectKey,
-                newExercise.getTestRepositoryUrl(), Optional.empty());
-
-        try {
-            continuousIntegrationService.get().triggerBuild(templateParticipation);
-            continuousIntegrationService.get().triggerBuild(solutionParticipation);
-        }
-        catch (HttpException e) {
-            log.error("Unable to trigger imported build plans", e);
-            throw e;
-        }
-    }
-
-    private void giveCIProjectPermissions(ProgrammingExercise exercise) {
+    public void giveCIProjectPermissions(ProgrammingExercise exercise) {
         final var instructorGroup = exercise.getCourse().getInstructorGroupName();
         final var teachingAssistantGroup = exercise.getCourse().getTeachingAssistantGroupName();
 
@@ -950,50 +708,6 @@ public class ProgrammingExerciseService {
         if (teachingAssistantGroup != null) {
             continuousIntegrationService.get().giveProjectPermissions(exercise.getProjectKey(), List.of(teachingAssistantGroup), List.of(CIPermission.READ));
         }
-    }
-
-    /**
-     * Remove the write permissions for all students for their programming exercise repository.
-     * They will still be able to read the code, but won't be able to change it.
-     *
-     * Requests are executed in batches so that the VCS is not overloaded with requests.
-     *
-     * @param programmingExerciseId     ProgrammingExercise id.
-     * @return a list of participations for which the locking operation has failed. If everything went as expected, this should be an empty list.
-     * @throws EntityNotFoundException  if the programming exercise can't be found.
-     */
-    public List<ProgrammingExerciseStudentParticipation> removeWritePermissionsFromAllStudentRepositories(Long programmingExerciseId) throws EntityNotFoundException {
-        log.info("Invoking scheduled task 'remove write permissions from all student repositories' for programming exercise with id " + programmingExerciseId + ".");
-
-        ProgrammingExercise programmingExercise = findByIdWithEagerStudentParticipations(programmingExerciseId);
-        List<ProgrammingExerciseStudentParticipation> failedLockOperations = new LinkedList<>();
-
-        int index = 0;
-        for (StudentParticipation studentParticipation : programmingExercise.getStudentParticipations()) {
-            // Execute requests in batches instead all at once.
-            if (index > 0 && index % EXTERNAL_SYSTEM_REQUEST_BATCH_SIZE == 0) {
-                try {
-                    log.info("Sleep for {}s during removeWritePermissionsFromAllStudentRepositories", EXTERNAL_SYSTEM_REQUEST_BATCH_WAIT_TIME_MS / 1000);
-                    Thread.sleep(EXTERNAL_SYSTEM_REQUEST_BATCH_WAIT_TIME_MS);
-                }
-                catch (InterruptedException ex) {
-                    log.error("Exception encountered when pausing before locking the student repositories for exercise " + programmingExerciseId, ex);
-                }
-            }
-
-            ProgrammingExerciseStudentParticipation programmingExerciseStudentParticipation = (ProgrammingExerciseStudentParticipation) studentParticipation;
-            try {
-                versionControlService.get().setRepositoryPermissionsToReadOnly(programmingExerciseStudentParticipation.getRepositoryUrlAsUrl(), programmingExercise.getProjectKey(),
-                        programmingExerciseStudentParticipation.getStudent().getLogin());
-            }
-            catch (Exception e) {
-                log.error("Removing write permissions failed for programming exercise with id " + programmingExerciseId + " for student repository with participation id "
-                        + studentParticipation.getId());
-                failedLockOperations.add(programmingExerciseStudentParticipation);
-            }
-            index++;
-        }
-        return failedLockOperations;
     }
 
     /**
@@ -1014,205 +728,6 @@ public class ProgrammingExerciseService {
     }
 
     /**
-     * Copied test cases from one exercise to another. The test cases will get new IDs, thus being saved as a new entity.
-     * The remaining contents stay the same, especially the weights.
-     *
-     * @param templateExercise The template exercise which test cases should get copied
-     * @param targetExercise The new exercise to which all test cases should get copied to
-     */
-    private void importTestCases(final ProgrammingExercise templateExercise, final ProgrammingExercise targetExercise) {
-        targetExercise.setTestCases(templateExercise.getTestCases().stream().map(testCase -> {
-            final var copy = new ProgrammingExerciseTestCase();
-
-            // Copy everything except for the referenced exercise
-            copy.setActive(testCase.isActive());
-            copy.setAfterDueDate(testCase.isAfterDueDate());
-            copy.setTestName(testCase.getTestName());
-            copy.setWeight(testCase.getWeight());
-            copy.setExercise(targetExercise);
-            programmingExerciseTestCaseRepository.save(copy);
-            return copy;
-        }).collect(Collectors.toSet()));
-    }
-
-    /**
-     * Sets up the test repository for a new exercise by setting the repository URL. This does not create the actual
-     * repository on the version control server!
-     *
-     * @param newExercise
-     * @param projectKey
-     */
-    private void setupTestRepository(ProgrammingExercise newExercise, String projectKey) {
-        final var testRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TESTS.getName();
-        newExercise.setTestRepositoryUrl(versionControlService.get().getCloneRepositoryUrl(projectKey, testRepoName).toString());
-    }
-
-    /**
-     * Sets up a new exercise for importing it by setting all values, that should either never get imported, or
-     * for which we should create new entities (e.g. test cases) to null. This ensures that we do not copy
-     * anything by accident.
-     *
-     * @param newExercise
-     */
-    private void setupExerciseForImport(ProgrammingExercise newExercise) {
-        newExercise.setId(null);
-        newExercise.setTemplateParticipation(null);
-        newExercise.setSolutionParticipation(null);
-        newExercise.setExerciseHints(null);
-        newExercise.setTestCases(null);
-        newExercise.setAttachments(null);
-        newExercise.setNumberOfMoreFeedbackRequests(null);
-        newExercise.setNumberOfComplaints(null);
-        newExercise.setNumberOfAssessments(null);
-        newExercise.setTutorParticipations(null);
-        newExercise.setExampleSubmissions(null);
-        newExercise.setStudentQuestions(null);
-        newExercise.setStudentParticipations(null);
-    }
-
-    /**
-     * Get participations of coding exercises of a requested list of students packed together in one zip file.
-     *
-     * @param programmingExerciseId the id of the exercise entity
-     * @param participations participations that should be exported
-     * @param repositoryExportOptions the options that should be used for the export
-     * @return a zip file containing all requested participations
-     */
-    public java.io.File exportStudentRepositories(Long programmingExerciseId, @NotNull List<ProgrammingExerciseStudentParticipation> participations,
-            RepositoryExportOptionsDTO repositoryExportOptions) {
-        // The downloaded repos should be cloned into another path in order to not interfere with the repo used by the student
-        String repoDownloadClonePath = REPO_DOWNLOAD_CLONE_PATH;
-
-        ProgrammingExercise programmingExercise = programmingExerciseRepository.findWithTemplateParticipationAndSolutionParticipationById(programmingExerciseId).get();
-
-        if (repositoryExportOptions.isExportAllStudents()) {
-            log.info("Request to export all student repositories of programming exercise " + programmingExerciseId + " with title '" + programmingExercise.getTitle() + "'");
-        }
-        else {
-            log.info("Request to export the repositories of programming exercise " + programmingExerciseId + " with title '" + programmingExercise.getTitle()
-                    + "' of the following students: " + participations.stream().map(p -> p.getStudent().getLogin()).collect(Collectors.joining(", ")));
-        }
-
-        List<Path> zippedRepoFiles = new ArrayList<>();
-        for (ProgrammingExerciseStudentParticipation participation : participations) {
-            Repository repo = null;
-            try {
-                if (participation.getRepositoryUrlAsUrl() == null) {
-                    log.warn("Ignore participation " + participation.getId() + " for export, because its repository URL is null");
-                    continue;
-                }
-                repo = gitService.getOrCheckoutRepository(participation, repoDownloadClonePath);
-                gitService.resetToOriginMaster(repo); // start with clean state
-
-                if (repositoryExportOptions.isFilterLateSubmissions() && repositoryExportOptions.getFilterLateSubmissionsDate() != null) {
-                    log.debug("Filter late submissions for participation {}", participation.toString());
-                    Optional<Submission> lastValidSubmission = participation.getSubmissions().stream()
-                            .filter(s -> s.getSubmissionDate() != null && s.getSubmissionDate().isBefore(repositoryExportOptions.getFilterLateSubmissionsDate()))
-                            .max(Comparator.comparing(Submission::getSubmissionDate));
-
-                    gitService.filterLateSubmissions(repo, lastValidSubmission, repositoryExportOptions.getFilterLateSubmissionsDate());
-                }
-
-                if (repositoryExportOptions.isAddStudentName()) {
-                    log.debug("Adding student name to participation {}", participation.toString());
-                    addStudentIdToProjectName(repo, programmingExercise, participation);
-                }
-
-                if (repositoryExportOptions.isCombineStudentCommits()) {
-                    log.debug("Combining commits for participation {}", participation.toString());
-                    gitService.combineAllStudentCommits(repo, programmingExercise);
-                }
-
-                if (repositoryExportOptions.isNormalizeCodeStyle()) {
-                    try {
-                        log.debug("Normalizing code style for participation {}", participation.toString());
-                        fileService.normalizeLineEndingsDirectory(repo.getLocalPath().toString());
-                        fileService.convertToUTF8Directory(repo.getLocalPath().toString());
-                    }
-                    catch (Exception ex) {
-                        log.warn("Cannot normalize code style in the repo " + repo.getLocalPath() + " due to the following exception: " + ex.getMessage());
-                    }
-                }
-
-                log.debug("Create temporary zip file for repository " + repo.getLocalPath().toString());
-                Path zippedRepoFile = gitService.zipRepository(repo, repoDownloadClonePath);
-                zippedRepoFiles.add(zippedRepoFile);
-            }
-            catch (IOException | GitException | GitAPIException | InterruptedException ex) {
-                log.error("export student repository " + participation.getRepositoryUrlAsUrl() + " in exercise '" + programmingExercise.getTitle() + "' did not work as expected: "
-                        + ex.getMessage());
-            }
-            finally {
-                // we do some cleanup here to prevent future errors with file handling
-                // We can always delete the repository as it won't be used by the student (separate path)
-                if (repo != null) {
-                    try {
-                        log.debug("Delete temporary repository " + repo.getLocalPath().toString());
-                        gitService.deleteLocalRepository(participation, repoDownloadClonePath);
-                    }
-                    catch (Exception ex) {
-                        log.warn("Could not delete temporary repository " + repo.getLocalPath().toString() + ": " + ex.getMessage());
-                    }
-                }
-            }
-        }
-        if (zippedRepoFiles.isEmpty()) {
-            log.warn("The zip file could not be created. Ignoring the request to export repositories for exercise " + programmingExercise.getTitle());
-            return null;
-        }
-        try {
-            // create a large zip file with all zipped repos and provide it for download
-            log.debug("Create zip file for all repositories");
-            Path zipFilePath = Paths.get(zippedRepoFiles.get(0).getParent().toString(),
-                    programmingExercise.getCourse().getShortName() + "-" + programmingExercise.getShortName() + "-" + System.currentTimeMillis() + ".zip");
-            createZipFile(zipFilePath, zippedRepoFiles);
-            scheduleForDeletion(zipFilePath, 15);
-            log.info("Export student repositories of programming exercise " + programmingExerciseId + " with title '" + programmingExercise.getTitle() + "' was successful.");
-            return new java.io.File(zipFilePath.toString());
-        }
-        catch (IOException ex) {
-            log.error("Export students repositories for exercise '" + programmingExercise.getTitle() + "' did not work as expected: " + ex.getMessage());
-        }
-        finally {
-            // we do some cleanup here to prevent future errors with file handling
-            log.debug("Delete all temporary zip repo files");
-            // delete the temporary zipped repo files
-            for (Path zippedRepoFile : zippedRepoFiles) {
-                try {
-                    Files.delete(zippedRepoFile);
-                }
-                catch (Exception ex) {
-                    log.warn("Could not delete file " + zippedRepoFile + ". Error message: " + ex.getMessage());
-                }
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Create a zipfile of the given paths and save it in the zipFilePath
-     *
-     * @param zipFilePath path where the zipfile should be saved
-     * @param paths the paths that should be zipped
-     * @throws IOException if an error occured while zipping
-     */
-    private void createZipFile(Path zipFilePath, List<Path> paths) throws IOException {
-        try (ZipOutputStream zipOutputStream = new ZipOutputStream(Files.newOutputStream(zipFilePath))) {
-            paths.stream().filter(path -> !Files.isDirectory(path)).forEach(path -> {
-                ZipEntry zipEntry = new ZipEntry(path.toString());
-                try {
-                    zipOutputStream.putNextEntry(zipEntry);
-                    Files.copy(path, zipOutputStream);
-                    zipOutputStream.closeEntry();
-                }
-                catch (Exception e) {
-                    log.error("Create zip file error", e);
-                }
-            });
-        }
-    }
-
-    /**
      * @param exerciseId the exercise we are interested in
      * @return the number of programming submissions which should be assessed
      */
@@ -1226,32 +741,5 @@ public class ProgrammingExerciseService {
      */
     public long countSubmissionsToAssessByCourseId(Long courseId) {
         return programmingExerciseRepository.countByCourseIdSubmittedBeforeDueDate(courseId);
-    }
-
-    private Map<Path, ScheduledFuture> futures = new HashMap<>();
-
-    private ScheduledExecutorService executor = Executors.newScheduledThreadPool(Runtime.getRuntime().availableProcessors());
-
-    private static final TimeUnit MINUTES = TimeUnit.MINUTES; // your time unit
-
-    /**
-     * Schedule the deletion of the given path with a given delay
-     *
-     * @param path The path that should be deleted
-     * @param delayInMinutes The delay in minutes after which the path should be deleted
-     */
-    private void scheduleForDeletion(Path path, long delayInMinutes) {
-        ScheduledFuture future = executor.schedule(() -> {
-            try {
-                log.info("Delete file " + path);
-                Files.delete(path);
-                futures.remove(path);
-            }
-            catch (IOException e) {
-                log.error("Deleting the file " + path + " did not work", e);
-            }
-        }, delayInMinutes, MINUTES);
-
-        futures.put(path, future);
     }
 }
