+++ src/main/java/de/tum/in/www1/artemis/web/rest/ProgrammingExerciseParticipationResource.java
@@ -147,7 +147,7 @@ public class ProgrammingExerciseParticipationResource {
     public ResponseEntity<Map<Long, Optional<ProgrammingSubmission>>> getLatestPendingSubmissionsByExerciseId(@PathVariable Long exerciseId) {
         ProgrammingExercise programmingExercise;
         try {
-            programmingExercise = programmingExerciseService.findById(exerciseId);
+            programmingExercise = programmingExerciseService.findWithTemplateParticipationAndSolutionParticipationById(exerciseId);
         }
         catch (EntityNotFoundException ex) {
             return notFound();
