+++ src/main/java/de/tum/in/www1/artemis/web/rest/ProgrammingExerciseResource.java
@@ -44,7 +44,6 @@ import de.tum.in.www1.artemis.service.connectors.ContinuousIntegrationService;
 import de.tum.in.www1.artemis.service.connectors.VersionControlService;
 import de.tum.in.www1.artemis.service.feature.Feature;
 import de.tum.in.www1.artemis.service.feature.FeatureToggle;
-import de.tum.in.www1.artemis.service.scheduled.ProgrammingExerciseScheduleService;
 import de.tum.in.www1.artemis.web.rest.dto.PageableSearchDTO;
 import de.tum.in.www1.artemis.web.rest.dto.RepositoryExportOptionsDTO;
 import de.tum.in.www1.artemis.web.rest.dto.SearchResultPageDTO;
@@ -81,11 +80,11 @@ public class ProgrammingExerciseResource {
 
     private final ProgrammingExerciseService programmingExerciseService;
 
-    private final ProgrammingExerciseScheduleService programmingExerciseScheduleService;
+    private final ProgrammingExerciseImportService programmingExerciseImportService;
+
+    private final ProgrammingExerciseExportService programmingExerciseExportService;
 
     private final StudentParticipationRepository studentParticipationRepository;
-
-    private final GroupNotificationService groupNotificationService;
 
     private final String packageNameRegex = "^[a-z][a-z0-9_]*(\\.[a-z0-9_]+)+[0-9a-z_]$";
 
@@ -93,8 +92,8 @@ public class ProgrammingExerciseResource {
 
     public ProgrammingExerciseResource(ProgrammingExerciseRepository programmingExerciseRepository, UserService userService, AuthorizationCheckService authCheckService,
             CourseService courseService, Optional<ContinuousIntegrationService> continuousIntegrationService, Optional<VersionControlService> versionControlService,
-            ExerciseService exerciseService, ProgrammingExerciseService programmingExerciseService, GroupNotificationService groupNotificationService,
-            ProgrammingExerciseScheduleService programmingExerciseScheduleService, StudentParticipationRepository studentParticipationRepository) {
+            ExerciseService exerciseService, ProgrammingExerciseService programmingExerciseService, StudentParticipationRepository studentParticipationRepository,
+            ProgrammingExerciseImportService programmingExerciseImportService, ProgrammingExerciseExportService programmingExerciseExportService) {
         this.programmingExerciseRepository = programmingExerciseRepository;
         this.userService = userService;
         this.courseService = courseService;
@@ -103,9 +102,9 @@ public class ProgrammingExerciseResource {
         this.versionControlService = versionControlService;
         this.exerciseService = exerciseService;
         this.programmingExerciseService = programmingExerciseService;
-        this.programmingExerciseScheduleService = programmingExerciseScheduleService;
         this.studentParticipationRepository = studentParticipationRepository;
-        this.groupNotificationService = groupNotificationService;
+        this.programmingExerciseImportService = programmingExerciseImportService;
+        this.programmingExerciseExportService = programmingExerciseExportService;
     }
 
     /**
@@ -136,49 +135,6 @@ public class ProgrammingExerciseResource {
                     .body(null);
         }
         return null;
-    }
-
-    /**
-     * POST /programming-exercises : Create a new programmingExercise.
-     *
-     * @param programmingExercise the programmingExercise to create
-     * @return the ResponseEntity with status 201 (Created) and with body the new programmingExercise, or with status 400 (Bad Request) if the programmingExercise has already an ID
-     * @throws URISyntaxException if the Location URI syntax is incorrect
-     */
-    @PostMapping(Endpoints.PROGRAMMING_EXERCISES)
-    @PreAuthorize("hasAnyRole('INSTRUCTOR', 'ADMIN')")
-    @FeatureToggle(Feature.PROGRAMMING_EXERCISES)
-    public ResponseEntity<ProgrammingExercise> createProgrammingExercise(@RequestBody ProgrammingExercise programmingExercise) throws URISyntaxException {
-        log.debug("REST request to save ProgrammingExercise : {}", programmingExercise);
-        if (programmingExercise.getId() != null) {
-            return ResponseEntity.badRequest()
-                    .headers(HeaderUtil.createFailureAlert(applicationName, true, ENTITY_NAME, "idexists", "A new programmingExercise cannot already have an ID")).body(null);
-        }
-        // fetch course from database to make sure client didn't change groups
-        Course course = courseService.findOne(programmingExercise.getCourse().getId());
-        User user = userService.getUserWithGroupsAndAuthorities();
-        if (!authCheckService.isInstructorInCourse(course, user) && !authCheckService.isAdmin()) {
-            return forbidden();
-        }
-
-        ResponseEntity<ProgrammingExercise> errorResponse = checkProgrammingExerciseForError(programmingExercise);
-        if (errorResponse != null) {
-            return errorResponse;
-        }
-
-        // We need to save the programming exercise BEFORE saving the participations to avoid transient state exceptions.
-        // This is only necessary for linked exercises, however we don't differentiate this with a separate endpoint.
-        programmingExercise.generateAndSetProjectKey();
-        programmingExercise = programmingExerciseRepository.save(programmingExercise);
-        // Only save after checking for errors
-        programmingExerciseService.saveParticipations(programmingExercise);
-
-        ProgrammingExercise savedProgrammingExercise = programmingExerciseRepository.save(programmingExercise);
-
-        programmingExerciseScheduleService.scheduleExerciseIfRequired(savedProgrammingExercise);
-        groupNotificationService.notifyTutorGroupAboutExerciseCreated(programmingExercise);
-        return ResponseEntity.created(new URI("/api/programming-exercises/" + savedProgrammingExercise.getId()))
-                .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, savedProgrammingExercise.getTitle())).body(savedProgrammingExercise);
     }
 
     /**
@@ -282,13 +238,9 @@ public class ProgrammingExerciseResource {
         }
 
         try {
-            ProgrammingExercise result = programmingExerciseService.setupProgrammingExercise(programmingExercise); // Setup all repositories etc
-
-            // TODO: should the call `scheduleExerciseIfRequired` not be moved into the service?
-            programmingExerciseScheduleService.scheduleExerciseIfRequired(result);
-            groupNotificationService.notifyTutorGroupAboutExerciseCreated(result);
-            return ResponseEntity.created(new URI("/api/programming-exercises" + result.getId()))
-                    .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getTitle())).body(result);
+            ProgrammingExercise newProgrammingExercise = programmingExerciseService.setupProgrammingExercise(programmingExercise); // Setup all repositories etc
+            return ResponseEntity.created(new URI("/api/programming-exercises" + newProgrammingExercise.getId()))
+                    .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, newProgrammingExercise.getTitle())).body(newProgrammingExercise);
         }
         catch (Exception e) {
             log.error("Error while setting up programming exercise", e);
@@ -302,11 +254,12 @@ public class ProgrammingExerciseResource {
      *
      * This will import the whole exercise, including all base build plans (template, solution) and repositories
      * (template, solution, test). Referenced entities, s.a. the test cases or the hints will get cloned and assigned
-     * a new id. For a concrete list of what gets copied and what not have a look at {@link ProgrammingExerciseService#importProgrammingExerciseBasis(ProgrammingExercise, ProgrammingExercise)}
+     * a new id. For a concrete list of what gets copied and what not have a look
+     * at {@link ProgrammingExerciseImportService#importProgrammingExerciseBasis(ProgrammingExercise, ProgrammingExercise)}
      *
-     * @see ProgrammingExerciseService#importProgrammingExerciseBasis(ProgrammingExercise, ProgrammingExercise)
-     * @see ProgrammingExerciseService#importBuildPlans(ProgrammingExercise, ProgrammingExercise)
-     * @see ProgrammingExerciseService#importRepositories(ProgrammingExercise, ProgrammingExercise)
+     * @see ProgrammingExerciseImportService#importProgrammingExerciseBasis(ProgrammingExercise, ProgrammingExercise)
+     * @see ProgrammingExerciseImportService#importBuildPlans(ProgrammingExercise, ProgrammingExercise)
+     * @see ProgrammingExerciseImportService#importRepositories(ProgrammingExercise, ProgrammingExercise)
      * @param sourceExerciseId The ID of the original exercise which should get imported
      * @param newExercise The new exercise containing values that should get overwritten in the imported exercise, s.a. the title or difficulty
      * @return The imported exercise (200), a not found error (404) if the template does not exist, or a forbidden error
@@ -334,13 +287,13 @@ public class ProgrammingExerciseResource {
         }
 
         final var originalProgrammingExercise = optionalOriginalProgrammingExercise.get();
-        final var importedProgrammingExercise = programmingExerciseService.importProgrammingExerciseBasis(originalProgrammingExercise, newExercise);
+        final var importedProgrammingExercise = programmingExerciseImportService.importProgrammingExerciseBasis(originalProgrammingExercise, newExercise);
         HttpHeaders responseHeaders;
-        programmingExerciseService.importRepositories(originalProgrammingExercise, importedProgrammingExercise);
+        programmingExerciseImportService.importRepositories(originalProgrammingExercise, importedProgrammingExercise);
         try {
             // TODO: We have removed the automatic build trigger from test to base for new programming exercises. We need to also remove this build trigger manually on the case of
             // an import as the source exercise might still have this trigger.
-            programmingExerciseService.importBuildPlans(originalProgrammingExercise, importedProgrammingExercise);
+            programmingExerciseImportService.importBuildPlans(originalProgrammingExercise, importedProgrammingExercise);
             responseHeaders = HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, importedProgrammingExercise.getTitle());
         }
         catch (HttpException e) {
@@ -372,7 +325,7 @@ public class ProgrammingExerciseResource {
             @RequestParam(value = "notificationText", required = false) String notificationText) throws URISyntaxException {
         log.debug("REST request to update ProgrammingExercise : {}", programmingExercise);
         if (programmingExercise.getId() == null) {
-            return createProgrammingExercise(programmingExercise);
+            return badRequest();
         }
         // fetch course from database to make sure client didn't change groups
         Course course = courseService.findOne(programmingExercise.getCourse().getId());
@@ -387,13 +340,7 @@ public class ProgrammingExerciseResource {
         }
 
         // Only save after checking for errors
-        ProgrammingExercise savedProgrammingExercise = programmingExerciseRepository.save(programmingExercise);
-
-        // TODO: should the call `scheduleExerciseIfRequired` not be moved into the service?
-        programmingExerciseScheduleService.scheduleExerciseIfRequired(savedProgrammingExercise);
-        if (notificationText != null) {
-            groupNotificationService.notifyStudentGroupAboutExerciseUpdate(savedProgrammingExercise, notificationText);
-        }
+        ProgrammingExercise savedProgrammingExercise = programmingExerciseService.updateProgrammingExercise(programmingExercise, notificationText);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, programmingExercise.getTitle())).body(savedProgrammingExercise);
     }
 
@@ -412,16 +359,13 @@ public class ProgrammingExerciseResource {
         log.debug("REST request to update ProgrammingExercise with new problem statement: {}", updatedProblemStatement);
         ProgrammingExercise updatedProgrammingExercise;
         try {
-            updatedProgrammingExercise = programmingExerciseService.updateProblemStatement(exerciseId, updatedProblemStatement);
+            updatedProgrammingExercise = programmingExerciseService.updateProblemStatement(exerciseId, updatedProblemStatement, notificationText);
         }
         catch (IllegalAccessException ex) {
             return forbidden();
         }
         catch (EntityNotFoundException ex) {
             return notFound();
-        }
-        if (notificationText != null) {
-            groupNotificationService.notifyStudentGroupAboutExerciseUpdate(updatedProgrammingExercise, notificationText);
         }
         return ResponseEntity.ok().headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, updatedProgrammingExercise.getTitle()))
                 .body(updatedProgrammingExercise);
@@ -670,7 +614,7 @@ public class ProgrammingExerciseResource {
                     .body(null);
         }
 
-        File zipFile = programmingExerciseService.exportStudentRepositories(exerciseId, exportedStudentParticipations, repositoryExportOptions);
+        File zipFile = programmingExerciseExportService.exportStudentRepositories(exerciseId, exportedStudentParticipations, repositoryExportOptions);
         if (zipFile == null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(applicationName, true, ENTITY_NAME, "internalServerError",
                     "There was an error on the server and the zip file could not be created.")).body(null);
